@@ -18,7 +18,7 @@
 		"compression": "^1.7.4",
 		"golden-fleece": "^1.0.9",
 		"highlight.js": "^10.5.0",
-		"marked": "^1.2.8",
+		"marked": "^2.0.0",
 		"polka": "^0.5.2",
 		"prismjs": "^1.23.0",
 		"sirv": "^1.0.11"
@@ -35,10 +35,10 @@
 		"@rollup/plugin-node-resolve": "^11.1.1",
 		"@rollup/plugin-replace": "^2.2.0",
 		"npm-run-all": "^4.1.5",
-		"rollup": "^2.38.4",
-		"rollup-plugin-svelte": "^7.0.0",
+		"rollup": "^2.41.0",
+		"rollup-plugin-svelte": "^7.1.0",
 		"rollup-plugin-terser": "^7.0.2",
-		"sapper": "^0.28.10",
-		"svelte": "^3.32.1"
+		"sapper": "^0.29.1",
+		"svelte": "^3.35.0"
 	}
 }
