@@ -1,6 +1,6 @@
+// eslint-disable-next-line import/no-unresolved
 import { files, shell, timestamp } from '@sapper/service-worker';
 
-// eslint-disable-next-line import/no-unresolved
 const ASSETS = `cache${timestamp}`;
 
 // `shell` is an array of all the files generated by the bundler,
